==> 2-CombineReleases.ipynb <==
## Step 2: Combine releases into a single feather file

* Read all Common Crawl releases into the same dataframe, and deduplicate based on URL and file contents

import os
import pandas as pd

# To parse GPX strings
import gpxpy

# For parallel_apply() in pandas
from pandarallel import pandarallel
pandarallel.initialize(progress_bar=True)

# To parse descriptions
from bs4 import BeautifulSoup

# To identify text language
import pycld2 as cld2 # pip install git+https://github.com/tmikus/pycld2.git

### Quick stats

dir = './interim/'

urls = (
    pd.concat([
        pd.read_feather(f'{dir}{r}') for r in os.listdir(dir) \
            if 'CC-MAIN-' in r and r.endswith('feather')
    ])
)

print('Total files:', urls.shape[0])
print('Downloaded files:', (~urls.responses.isna()).sum())
print('Downloaded files %:', round((~urls.responses.isna()).sum() / urls.shape[0] * 100, 1))
print('Unique URLs among downloaded:', urls.query('~responses.isna()').url.nunique())
print('Unique URLs among downloaded %:', round(urls.query('~responses.isna()').url.nunique() / (~urls.responses.isna()).sum() * 100, 1))

### Combine releases & deduplicate

dir = './interim/'

urls = (
    pd.concat([
        pd.read_feather(f'{dir}{r}') for r in os.listdir(dir) \
            if 'CC-MAIN-' in r and r.endswith('feather')
    ])
    .drop_duplicates(subset=['responses'])
    .drop_duplicates(subset=['url'])
    .reset_index()
    .drop(columns=['index'])
)

print('Number of files after deduplication:', urls.shape[0])

### Convert respones to GPX objects, and extract metadata

def get_gpx_metadata(s):
    try:
        gpx = gpxpy.parse(s)
        md = gpx.get_moving_data()

        # How many points are there in the first track?
        pts_in_first_track = sum(
            [ len(segment.points) for track in gpx.tracks[:1] for segment in track.segments ]
        )

        # What is the average number of points per 100 m in first track?
        pts_per_100m = pts_in_first_track / gpx.tracks[0].length_2d() * 100
        
        return {
            'n_tracks': len(gpx.tracks),
            'pts_per_100m': pts_per_100m,
            'description': gpx.description,
            'length_2d': round(gpx.length_2d())
        }
    except:
        return None

%%time
urls['gpx_metadata'] = urls.responses.parallel_apply(get_gpx_metadata)

urls['length_2d'] = urls.gpx_metadata.str['length_2d']
urls['pts_per_100m'] = urls.gpx_metadata.str['pts_per_100m']
urls['n_tracks'] = urls.gpx_metadata.str['n_tracks']

### Clean up the descriptions

%%capture

def clean_description(desc):
    """
    Returns a clean text description with HTML tags removed
    """
    try:
        desc = BeautifulSoup(desc).get_text()
        desc = desc.replace('\n', ' ').replace('\t', ' ').replace('\xa0', ' ')
        desc = desc.replace("l\'", ' ')
        desc = ' '.join(desc.strip().split())
        return desc
    except:
        return None

urls['description'] = urls.gpx_metadata.str['description'].apply(clean_description)

# Remove contents in square brackets using regex
urls.description = urls.description.str.replace(r'\[.*?\]', '', regex=True)

# Remove contents in curly brackets using regex
urls.description = urls.description.str.replace(r'\{.*?\}', '', regex=True)

# Strip trailing spaces
urls.description = urls.description.str.strip()

### Establish relevant description lengths

# We want descriptions of over 50 characters
# and ignore those that are too long (over 99th percentile ~ 2000 characters)
print(
    'Length at 99.0, 99.1, 99.2 percentiles: ',
    round(urls.description.str.len().quantile(0.990)),
    round(urls.description.str.len().quantile(0.991)),
    round(urls.description.str.len().quantile(0.992))
)

cap_value = 2000
urls['is_desc_long_enough'] = (
    urls.description.str.len().ge(50)
    & urls.description.str.len().le(cap_value)
)

### Identify source description language
* We want to exclude tracks whose description language is Unknown by `pycld2` (these are typically low quality)

def detect_language(s):
    try:
        _, _, lang = cld2.detect(str(s))
        return lang[0][1]
    except:
        return 'un' # Unknown language

urls['description_lang'] = 'un' # Set all to Unknown first

# For descriptions of appropriate length, guess language
idx = urls.is_desc_long_enough
urls.loc[idx , 'description_lang'] = urls.loc[idx, 'description'].apply(detect_language)

urls.description_lang.value_counts()

### Keep only relevant GPX files, and save to `interim/relevant_subset.feather`
* We want tracks with descriptions between 50-2000 characters
* With only 1 track
* Length is between 0.5km - 100km
* Description language is not *Unknown*
* At least 1 GPS point per 100 metres

relevant_subset = (
    urls
    .query('is_desc_long_enough')
    .query('n_tracks.eq(1)')
    .query('pts_per_100m.ge(1.0)')
    .query('length_2d.ge(500) & length_2d.le(100000)')
    .query('length_2d.ge(500) & length_2d.le(100000)')
    .query('description_lang.ne("un")')
    .reset_index()
    .drop(columns=['index'])
)

relevant_subset

relevant_subset.to_feather('interim/relevant_subset.feather')
