==> 3-QaWithLLM.ipynb <==
## Step 3: QA our subset with an LLM

* We let the LLM decide if the description is of an activity (hiking, cycling, etc)
* We use **Llama-3 8B Instruct** run locally with Ollama (visit https://ollama.com/ to install)
* Once installed, run `ollama run llama3` in Terminal

import pandas as pd
import requests
import json

# For progress_apply() in pandas
from tqdm import tqdm
tqdm.pandas()

relevant_subset = pd.read_feather('./interim/relevant_subset.feather')

def ask_llm(prompt, limit=9999):
    """
    Sends the `prompt` to the LLM running locally via Ollama.
    Temperature = 0 to get consistent output between runs.
    Works much faster for shorter limits, so consider
    setting to 1 for True/False queries
    """
    response = requests.post(
        url='http://localhost:11434/api/generate',
        json={
            'model': 'llama3',
            'prompt': prompt,
            'stream': False,
            'options': {
                'temperature': 0,
                'seed': 42,
                'num_predict': limit
            }
        }
    )
    
    return json.loads(response.text).get('response').strip()

def is_activity_description(text):
    prompt = f"""
    Does the text in triple quotes represent a high-quality and insightful
    route or track description, or an activity description such as hiking,
    cycling, or racing? Respond with `True` or `False`.
    If you are unsure, say `False`.
    Text: ```{text}```
    """
    return ask_llm(prompt, 1)


def contains_personal_info(text):
    prompt = f"""
    Does the text in triple quotes contain any personal information,
    such as someone's address or name? Respond with `True` or `False`.
    If you are unsure, say `True`.
    Text: ```{text}```
    """
    return ask_llm(prompt, 1)

### Determine if GPX description resembles an activity description
We only want relevant descriptions that actually describe the activity

relevant_subset['is_activity_description'] = (
    relevant_subset.description.progress_apply(is_activity_description)
)

#### Let's look at what the LLM has decided

(
    relevant_subset
    .query('is_activity_description == "True" and description_lang == "en"')
    .sample(5)
    .description
    .to_list()
)

is_activity_description('Note check opening times of Manor Park Cremitorium')

relevant_subset.is_activity_description.value_counts()

# Save just in case
relevant_subset.to_feather('./interim/relevant_subset_llm_checkpoint.feather')

### For those that resemble an activity, check for personal info
We don't want any descriptions with personal information

idx = relevant_subset.is_activity_description.eq('True')

relevant_subset.loc[idx, 'contains_personal_info'] = relevant_subset\
    .loc[idx, 'description'].progress_apply(contains_personal_info)

relevant_subset.contains_personal_info.value_counts()

# Save just in case
relevant_subset.to_feather('./interim/relevant_subset_llm_checkpoint.feather')

### 💾 Save final
We reset index because feather wants it default

(relevant_subset
 .to_feather('./interim/relevant_subset_llm.feather')
)
