==> 4-ProduceFinalDataset.ipynb <==
## Step 4: Produce final dataset 

### Read GPX files into a geodataframe
* First, write response strings into .gpx files
* Then, read them one by one and append into a list
* Construct a new geodataframe from the `relevant_subset` + these gpx geometries

import pandas as pd

# For all things geospatial
import geopandas as gpd
import shapely
from shapely.geometry import LineString, MultiLineString
from geopy.distance import geodesic
import srtm # pip install git+https://github.com/tkrajina/srtm.py
import gpxpy

# To mask phone numbers
import phonenumbers

# For parallel_apply() in pandas
from pandarallel import pandarallel
pandarallel.initialize(progress_bar=True)

# For translation
import argostranslate.package
import argostranslate.translate

relevant_subset = (
    pd.read_feather(
        './interim/relevant_subset_llm.feather'
    )
    .query('is_activity_description.eq("True")')
    .query('contains_personal_info.eq("False")')
)

relevant_subset.shape

### Get elevation data
* Where elevation data does not exist, use SRTM (see [SRTM.py](https://github.com/tkrajina/srtm.py) on GitHub)

relevant_subset.head(5)

geoms = []
elev_profiles = []
length_3d = []

for i, response in relevant_subset.responses.items():

    if i % 25 == 0:
        print(f'Processing {i}...')
    
    track = gpxpy.parse(response)
    elev_source = 'GPS'

    # If elevation does not exist, use SRTM
    if track.get_uphill_downhill().uphill == 0 and track.get_uphill_downhill().downhill == 0:
        elevation_data.add_elevations(track, smooth=True) # Cached, so will work fast 2nd/3rd/... time
        elev_source = 'DEM'

    length_3d.append( track.length_3d() )
    
    path = f'./gpx/{i}.gpx'
    
    with open(path, 'w') as f:
        f.write(track.to_xml())

    try:
        g = gpd.read_file(path, layer='tracks').iloc[0].geometry
        g = MultiLineString(g) if g.geom_type == 'LineString' else g
        elevations = gpd.read_file(path, layer='track_points').ele.tolist()

        # Add elevations to each point
        lines = []
        for line in g.geoms:
            lines.append( [ (x, y, elevations.pop(0)) for x, y in line.coords] )

        # Recreate a MultiLineString Z (with elevations)
        geoms.append(MultiLineString(lines))

        # Add elevation profile
        elev_profiles.append({
            'highest': track.get_elevation_extremes().maximum,
            'lowest': track.get_elevation_extremes().minimum,
            'uphill': track.get_uphill_downhill().uphill,
            'downhill': track.get_uphill_downhill().downhill,
            'source': elev_source
        })
        
    except:
        geoms.append(None)
        elev_profiles.append({})

assert len(elev_profiles) == len(geoms) == relevant_subset.shape[0]

final = (
    gpd.GeoDataFrame(
        relevant_subset, 
        geometry=geoms,
        crs=4326
    )
    .assign(
        elev_highest=[ x.get('highest', None) for x in elev_profiles ],
        elev_lowest=[ x.get('lowest', None) for x in elev_profiles ],
        uphill=[ x.get('uphill', None) for x in elev_profiles ],
        downhill=[ x.get('downhill', None) for x in elev_profiles ],
        elev_source= [ x.get('source', None) for x in elev_profiles ],
        length_3d=[round(x) for x in length_3d]
    )
    .dropna(subset=['geometry'])
)

# Convert 3-dimensional length from float to int
final.length_2d = final.length_2d.astype(int)

final

### Translate descriptions to English

# Only translate popular languages (5+ examples)
lang_counts = final.description_lang.value_counts()
relevant_langs = lang_counts[lang_counts.gt(5)].index.tolist()
relevant_langs.remove('hr') # Croatian does not exist in argos
relevant_langs

# Download and install Argos Translate packages
argostranslate.package.update_package_index()
available_packages = argostranslate.package.get_available_packages()

for lang in relevant_langs:
    if lang == 'en': # We don't want to translate from English
        continue

    try:
        package_to_install = next(
            filter(
                lambda x: x.from_code == lang and x.to_code == 'en', available_packages
            )
        )
        argostranslate.package.install_from_path(package_to_install.download())
    except:
        print(f'Could not install {lang}...')

%%time

def translate(text, lang_from, lang_to='en'):
    if lang_from == lang_to: # no need to translate en<->en
        return text

    if not lang_from in relevant_langs:
        return None
        
    return argostranslate.translate.translate(text, lang_from, lang_to)

final['description_en'] = final.parallel_apply(
    lambda row: translate(row['description'], row['description_lang']),
    axis=1
)

final.dropna(subset=['description_en'], inplace=True)

### Mask phone numbers, URLs, and emails in descriptions 🎭

# Remove emails - for regex, see https://html.spec.whatwg.org/multipage/input.html#valid-e-mail-address
email_regex = r"[a-zA-Z0-9.!#$%&'*+\/=?^_`{|}~-]+@[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?(?:\.[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?)*"

final.description = final.description.str.replace(email_regex, '<EMAIL>', regex=True)
final.description_en = final.description_en.str.replace(email_regex, '<EMAIL>', regex=True)

# Remove URLs - courtesy of https://stackoverflow.com/a/8234912/4361039
url_regex = r'((([A-Za-z]{3,9}:(?:\/\/)?)(?:[-;:&=\+\$,\w]+@)?[A-Za-z0-9.-]+(:[0-9]+)?|(?:www.|[-;:&=\+\$,\w]+@)[A-Za-z0-9.-]+)((?:\/[\+~%\/.\w\-_]*)?\??(?:[-\+=&;%@.\w_]*)#?(?:[\w]*))?)'

final.description = final.description.str.replace(url_regex, '<URL>', regex=True)
final.description_en = final.description_en.str.replace(url_regex, '<URL>', regex=True)

def mask_telephone(s):
    s = str(s)
    for match in phonenumbers.PhoneNumberMatcher(s, None):
        s = s.replace(match.raw_string, '<TELEPHONE>')
    return s

final.description = final.description.apply(mask_telephone)
final.description_en = final.description_en.apply(mask_telephone)

### Add country name
* Use country boundaries from https://www.geoboundaries.org/globalDownloads.html
* Assign country based on the very first point in the linestring

countries_gdf = gpd.read_file(
    'https://github.com/wmgeolab/geoBoundaries/raw/main/releaseData/CGAZ/geoBoundariesCGAZ_ADM0.gpkg'
).rename(columns={'shapeName': 'country'}).to_crs(4326)

start_points = final.geometry.apply(
    lambda g: shapely.get_point(g.geoms[0], 0)
).to_frame()

end_points = final.geometry.apply(
    lambda g: shapely.get_point(g.geoms[-1], -1)
).to_frame()

countries = gpd.sjoin_nearest(
    start_points, countries_gdf, how='left'
).country

final.loc[:, 'country'] = countries

### Measure distances between start and end points to determine if routes are circular

distances = [ geodesic(p1, p2).m for p1, p2 in list(zip(
    start_points.geometry.apply(lambda p: p.coords[0][:2][::-1]),
    end_points.geometry.apply(lambda p: p.coords[0][:2][::-1])
)) ]

is_circular = [ True if d <= 300 else False for d in distances]

final['is_circular'] = is_circular

### 💾 Finally, save to geopackage (.gpkg) 

(final
 .to_file('final/final.full.gpkg')
)

(final
 .filter([
     'url',
     'warc_filename',
     'warc_record_offset',
     'warc_record_length',
     
     'country',
     'description',
     'description_lang',
     'description_en',

     'elev_source',
     'elev_highest',
     'elev_lowest',
     'uphill',
     'downhill',
     
     'length_2d',
     'length_3d',
     'is_circular',
     
     'geometry'
 ])
 .rename(columns={
     'description': 'desc',
     'description_lang': 'desc_lang',
     'description_en': 'desc_en',
     'warc_filename': 'warc_file',
     'warc_record_offset': 'warc_offset',
     'warc_record_length': 'warc_len',
 })
 .to_file('final/final.gpkg')
)

final.elev_source.eq('GPS').mean()
